+++ R/box_package_usage_helper_functions.R
@@ -181,7 +181,7 @@ get_attached_pkg_functions <- function(xml) {
     list_output
   })
 
-  attached_functions$text <- unlist(aliases)[attached_functions$text]
+  attached_functions$text <- unlist(aliases)
   list(
     xml = attached_functions$xml_nodes,
     text = attached_functions$text